+++ app/src/main/java/vasyl/v/stoliarchuk/addresstracker/data/datasource/remote/retrofit/RestApi.java
@@ -4,13 +4,12 @@ import io.reactivex.Maybe;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
 import vasyl.v.stoliarchuk.addresstracker.data.datasource.remote.retrofit.entity.RetrofitPlace;
-import vasyl.v.stoliarchuk.addresstracker.data.entity.Place;
 
 import static vasyl.v.stoliarchuk.addresstracker.data.datasource.remote.retrofit.RestApi.RestConfig.GET_PLACE;
 
 public interface RestApi {
     @GET(GET_PLACE)
-    Maybe<RetrofitPlace> getPlace(@Query(value = "lat") float lat, @Query(value = "lon") float lon);
+    Maybe<RetrofitPlace> getPlace(@Query(value = "lat") double lat, @Query(value = "lon") double lon);
 
     interface RestConfig {
         String BASE_URL = "https://nominatim.openstreetmap.org";
